chore: replace ESLint & Prettier with Rome (#738)
* chore: replace eslint & prettier with Rome

* feat(ci): run rome format check in CI pipeline

* fix(ci): rome github action package version

* fix(ci): rome github action package version to v0.4

* feat(ci): disable rome format in test pipeline

* feat(ci): enable rome format in test pipeline

* fix(ci): rome CI command in test pipeline

* feat(refactor): run rome format script

* feat(refactor): format the code

* feat(refactor): apply safe & suggested fixes from linter

* feat(refactor): format the code

* chore: ignore theme build directory in rome.json

+++ packages/server/database/utils.js
@@ -4,56 +4,53 @@ const { v4: uuidv4 } = require("uuid");
 const logger = require("../utils/logger");
 
 const permissionExists = async (database, permission) => {
-	return await database
-		.select()
-		.from("permissions")
-		.where({
-			...permission
-		})
-		.first();
+  return await database
+    .select()
+    .from("permissions")
+    .where({
+      ...permission,
+    })
+    .first();
 };
 
 exports.addPermission = (database, permissions) => {
-	return permissions.forEach(async permission => {
-		const type = permission.split(":")[0];
-		const action = permission.split(":")[1];
+  return permissions.forEach(async (permission) => {
+    const type = permission.split(":")[0];
+    const action = permission.split(":")[1];
 
-		const exists = await permissionExists(database, {
-			type,
-			action
-		});
+    const exists = await permissionExists(database, {
+      type,
+      action,
+    });
 
-		if (exists) {
-			return logger.warn({
-				message: `Permssion ${type}:${action} already added`
-			});
-		}
+    if (exists) {
+      return logger.warn({
+        message: `Permission ${type}:${action} already added`,
+      });
+    }
 
-		await database
-			.insert({
-				id: uuidv4(),
-				type,
-				action
-			})
-			.into("permissions");
+    await database
+      .insert({
+        id: uuidv4(),
+        type,
+        action,
+      })
+      .into("permissions");
 
-		logger.info(`Adding permssion ${type}:${action}`);
-	});
+    logger.info(`Adding permission ${type}:${action}`);
+  });
 };
 
 exports.removePermission = async (database, permissions) => {
-	return await permissions.forEach(async permission => {
-		const type = permission.split(":")[0];
-		const action = permission.split(":")[1];
+  return await permissions.forEach(async (permission) => {
+    const type = permission.split(":")[0];
+    const action = permission.split(":")[1];
 
-		await database
-			.delete()
-			.from("permissions")
-			.where({
-				type,
-				action
-			});
+    await database.delete().from("permissions").where({
+      type,
+      action,
+    });
 
-		logger.info(`Removing permssion ${type}:${action}`);
-	});
+    logger.info(`Removing permission ${type}:${action}`);
+  });
 };
